requirements.txt
GitPython
requests
pathlib
